server/requirements.txt
livekit-agents
python-dotenv
livekit-plugins-deepgram
livekit-plugins-silero
python-dotenv
livekit-plugins-deepgram
livekit-plugins-silero
google-generativeai